+++ R/HermesData-methods.R
@@ -26,8 +26,8 @@
 #' @seealso [`cbind`] to column bind objects.
 #'
 #' @examples
-#' a <- HermesData(summarized_experiment[1:2542, ])
-#' b <- HermesData(summarized_experiment[2543:5085, ])
+#' a <- hermes_data[1:2542, ]
+#' b <- hermes_data[2543:5085, ]
 #' result <- rbind(a, b)
 #' class(result)
 NULL
@@ -59,8 +59,8 @@ NULL
 #' @seealso [`rbind`] to row bind objects.
 #'
 #' @examples
-#' a <- HermesData(summarized_experiment[, 1:10])
-#' b <- HermesData(summarized_experiment[, 11:20])
+#' a <- hermes_data[, 1:10]
+#' b <- hermes_data[, 11:20]
 #' result <- cbind(a, b)
 #' class(result)
 NULL
@@ -87,7 +87,7 @@ NULL
 #' @export `metadata<-`
 #'
 #' @examples
-#' a <- HermesData(summarized_experiment)
+#' a <- hermes_data
 #' metadata(a)
 #' metadata(a) <- list(new = "my metadata")
 #' metadata(a)
@@ -108,20 +108,16 @@ NULL
 #' @param ... not used.
 #'
 #' @return The [`S4Vectors::DataFrame`] with the gene annotations:
-#'   - `HGNC`
-#'   - `HGNCGeneName`
-#'   - `Chromosome`
-#'   - `StartBP`
-#'   - `EndBP`
-#'   - `WidthBP`
-#'   - `CanonicalTranscript`
-#'   - `ProteinTranscript`
+#'   - `symbol`
+#'   - `desc`
+#'   - `chromosome`
+#'   - `size`
 #'
 #' @importFrom BiocGenerics annotation
 #' @export
 #'
 #' @examples
-#' object <- HermesData(summarized_experiment)
+#' object <- hermes_data
 #' head(annotation(object))
 setMethod(
   f = "annotation",
@@ -136,21 +132,19 @@ setMethod(
 #'   character vector `.row_data_annotation_cols`.
 #' @export
 .row_data_annotation_cols <- c(
-  "HGNC",
-  "HGNCGeneName",
-  "Chromosome",
-  "StartBP",
-  "EndBP",
-  "WidthBP",
-  "CanonicalTranscript",
-  "ProteinTranscript"
+  "symbol",
+  "desc",
+  "chromosome",
+  "size"
 )
 
 #' @param value (`DataFrame`)\cr what should the annotations be replaced with.
 #'
-#' @note When trying to replace the annotation with completely missing values for any genes,
-#'   a warning will be given and the corresponding gene IDs will be saved in the
-#'   attribute `annotation.missing.genes`.
+#' @note When trying to replace the required annotations with completely missing
+#'   values for any genes, a warning will be given and the corresponding gene
+#'   IDs will be saved in the attribute `annotation.missing.genes`. Note also
+#'   that additional annotations beyond the required ones may be supplied and
+#'   will be stored.
 #'
 #' @importFrom BiocGenerics `annotation<-`
 #' @rdname annotation
@@ -161,17 +155,21 @@ setReplaceMethod(
   definition = function(object, value) {
     assert_that(
       identical(rownames(object), rownames(value)),
-      setequal(.row_data_annotation_cols, colnames(value))
+      all(.row_data_annotation_cols %in% colnames(value))
     )
-    row_is_all_na <- apply(X = value, MARGIN = 1L, FUN = all_na)
+    row_is_all_na <- apply(X = value[, .row_data_annotation_cols], MARGIN = 1L, FUN = all_na)
     if (any(row_is_all_na)) {
       warning(paste(
-        "annotations completely missing for", sum(row_is_all_na), "genes,",
+        "required annotations completely missing for", sum(row_is_all_na), "genes,",
         "see attribute `annotation.missing.genes` for the corresponding gene IDs"
       ))
       attr(object, "annotation.missing.genes") <- names(which(row_is_all_na))
     }
     rowData(object)[, .row_data_annotation_cols] <- value[, .row_data_annotation_cols]
+    add_annotation_cols <- setdiff(colnames(value), .row_data_annotation_cols)
+    if (length(add_annotation_cols)) {
+      rowData(object)[, add_annotation_cols] <- value[, add_annotation_cols, drop = FALSE]
+    }
     validObject(object)
     object
   }
@@ -197,7 +195,7 @@ setReplaceMethod(
 #' @export
 #'
 #' @examples
-#' a <- HermesData(summarized_experiment)
+#' a <- hermes_data
 #' result <- counts(a)
 #' class(result)
 #' head(result)
@@ -247,7 +245,7 @@ setReplaceMethod(
 #' @export
 #'
 #' @examples
-#' a <- HermesData(summarized_experiment)
+#' a <- hermes_data
 #' prefix(a)
 setGeneric("prefix", def = function(object, ...) {
   object@prefix
@@ -274,7 +272,7 @@ setGeneric("genes", def = function(object) standardGeneric("genes"))
 #' @rdname genes
 #' @export
 #' @examples
-#' a <- HermesData(summarized_experiment)
+#' a <- hermes_data
 #' genes(a)
 setMethod(
   f = "genes",
@@ -305,7 +303,7 @@ setMethod(
 #' @importFrom Biobase samples
 #' @export
 #' @examples
-#' a <- HermesData(summarized_experiment)
+#' a <- hermes_data
 #' samples(a)
 setMethod(
   f = "samples",
@@ -338,14 +336,14 @@ setMethod(
 #' @return The subsetted [`AnyHermesData`] object.
 #'
 #' @examples
-#' a <- HermesData(summarized_experiment)
+#' a <- hermes_data
 #' a
 #'
 #' # Subset both genes and samples.
-#' subset(a, subset = LowExpressionFlag, select = DISCSTUD == "N")
+#' subset(a, subset = low_expression_flag, select = DISCSTUD == "N")
 #'
 #' # Subset only genes.
-#' subset(a, subset = Chromosome == "2")
+#' subset(a, subset = chromosome == "2")
 #'
 #' # Subset only samples.
 #'subset(a, select = AGE > 18)
@@ -383,11 +381,11 @@ setGeneric("filter", function(object, ...) standardGeneric("filter"))
 #' @export
 #'
 #' @examples
-#' object <- HermesData(summarized_experiment)
-#' result <- h_has_req_annotations(object, "WidthBP")
+#' object <- hermes_data
+#' result <- h_has_req_annotations(object, "size")
 #' all(result)
-#' rowData(object)$WidthBP[1] <- NA # nolint
-#' which(!h_has_req_annotations(object, "WidthBP"))
+#' rowData(object)$size[1] <- NA # nolint
+#' which(!h_has_req_annotations(object, "size"))
 h_has_req_annotations <- function(object,
                                   annotation_required) {
   assert_that(
@@ -401,11 +399,11 @@ h_has_req_annotations <- function(object,
 #' @rdname filter
 #'
 #' @details
-#' - Only genes without low expression (`LowExpressionFlag`) and samples
-#'   without low depth (`LowDepthFlag`) or technical failure (`TechnicalFailureFlag`)
+#' - Only genes without low expression (`low_expression_flag`) and samples
+#'   without low depth (`low_depth_flag`) or technical failure (`tech_failure_flag`)
 #'   remain in the returned filtered object.
 #' - Also required gene annotation columns can be specified, so that genes which are not complete
-#'   for these columns are filtered out. By default this is the `WidthBP` column, which is needed
+#'   for these columns are filtered out. By default this is the `size` column, which is needed
 #'   for default normalization of the object.
 #'
 #' @param what (`character`)\cr specify whether to apply the filter on `genes` and / or `samples`.
@@ -417,7 +415,7 @@ h_has_req_annotations <- function(object,
 #'
 #' @export
 #' @examples
-#' a <- HermesData(summarized_experiment)
+#' a <- hermes_data
 #' dim(a)
 #'
 #' # Filter genes and samples on default QC flags.
@@ -431,13 +429,13 @@ h_has_req_annotations <- function(object,
 #' result <- filter(a, what = "samples")
 #'
 #' # Filter only genes, and require certain annotations to be present.
-#' result <- filter(a, what = "genes", annotation_required = c("StartBP", "EndBP", "WidthBP"))
+#' result <- filter(a, what = "genes", annotation_required = c("size"))
 setMethod(
   f = "filter",
   signature = signature(object = "AnyHermesData"),
   definition = function(object,
                         what = c("genes", "samples"),
-                        annotation_required = "WidthBP") {
+                        annotation_required = "size") {
     low_exp <- get_low_expression(object)
     low_depth <- get_low_depth(object)
     tech_fail <- get_tech_failure(object)
@@ -485,7 +483,7 @@ setMethod(
 #'   `colData()` or `rowData()`.
 #'
 #' @examples
-#' object <- HermesData(summarized_experiment)
+#' object <- hermes_data
 NULL
 
 # extraColDataNames ----
@@ -567,7 +565,7 @@ setGeneric("summary")
 #' @export
 #'
 #' @examples
-#' object <- HermesData(summarized_experiment)
+#' object <- hermes_data
 #' object_summary <- summary(object)
 #'
 #' # We can access parts of this S4 object with the slot operator.
@@ -712,7 +710,7 @@ setMethod(
 #' @export
 #'
 #' @examples
-#' object <- HermesData(summarized_experiment)
+#' object <- hermes_data
 #' object
 setMethod(
   f = "show",