+++ R/draw_barplot.R
@@ -20,7 +20,7 @@
 #' @export
 #'
 #' @examples
-#' object <- HermesData(summarized_experiment) %>%
+#' object <- hermes_data %>%
 #'   add_quality_flags() %>%
 #'   filter() %>%
 #'   normalize()