+++ R/pca_cor_samplevar.R
@@ -20,7 +20,7 @@ NULL
 #' @export
 #'
 #' @examples
-#' object <- HermesData(summarized_experiment) %>%
+#' object <- hermes_data %>%
 #'   add_quality_flags() %>%
 #'   filter() %>%
 #'   normalize()
@@ -29,7 +29,7 @@ NULL
 #' pca <- calc_pca(object)$x
 #'
 #' # Obtain the sample variable.
-#' x <- colData(object)$LowDepthFlag
+#' x <- colData(object)$AGE18
 #'
 #' # Correlate them.
 #' r2 <- h_pca_var_rsquared(pca, x)
@@ -82,7 +82,7 @@ h_pca_var_rsquared <- function(pca, x) {
 #' @export
 #'
 #' @examples
-#' object <- HermesData(summarized_experiment) %>%
+#' object <- hermes_data %>%
 #'   add_quality_flags() %>%
 #'   filter() %>%
 #'   normalize()
@@ -159,7 +159,7 @@ h_pca_df_r2_matrix <- function(pca, df) {
 #' @export
 #'
 #' @examples
-#' object <- HermesData(summarized_experiment) %>%
+#' object <- hermes_data %>%
 #'   add_quality_flags() %>%
 #'   filter() %>%
 #'   normalize()